+++ compiler/reader.go
@@ -78,8 +78,17 @@ func GetInfoCache() map[string]interface{} {
 	return infoCache
 }
 
+func ClearFileCache() {
+	fileCache = make(map[string][]byte, 0)
+}
+
 func ClearInfoCache() {
 	infoCache = make(map[string]interface{})
+}
+
+func ClearCaches() {
+	ClearFileCache()
+	ClearInfoCache()
 }
 
 // FetchFile gets a specified file from the local filesystem or a remote location.
