+++ testdata/README.md
@@ -1,3 +1,3 @@
-# test
+# testdata
 
 This directory contains test data used to verify gnostic.
