+++ testdata/errors/petstore-badproperties.errors
