+++ testdata/errors/petstore-missingversion.errors
