+++ testdata/errors/petstore-unresolvedrefs.errors
