+++ testdata/library-example-with-ext.json
