+++ testdata/library-example-with-ext.text.out
