+++ testdata/v2.0/yaml/petstore-separate/spec/swagger.text
