+++ testdata/v2.0/yaml/sample-petstore.out
